requirements.txt
aioice==0.10.1
aiortc==1.14.0
altair==5.5.0
attrs==25.4.0
av==16.0.1
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
deep-sort-realtime==1.3.2
defusedxml==0.7.1
dnspython==2.8.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-crc32c==1.7.1
idna==3.11
ifaddr==0.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
narwhals==2.9.0
networkx==3.3
numpy==2.1.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
polars==1.34.0
polars-runtime-32==1.34.0
protobuf==6.33.0
psutil==7.1.2
pyarrow==22.0.0
pycparser==2.23
pydeck==0.9.1
pyee==13.0.0
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
streamlit-webrtc==0.63.11
supervision==0.26.1
sympy==1.13.1
tenacity==9.1.2
toml==0.10.2
torch
torchaudio
torchvision
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.221
ultralytics-thop==2.0.17
urllib3==2.5.0
watchdog==6.0.0
# API and Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
# Testing
pytest==8.3.3
pytest-cov==6.0.0
pytest-asyncio==0.24.0
httpx==0.27.2
# Data Visualization
plotly==5.24.1
# Additional utilities
python-dotenv==1.0.1
# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Background jobs
celery==5.4.0
redis==5.2.0
# Monitoring
prometheus-client==0.21.0
